+++ test/src/test_json.cpp
@@ -325,3 +325,19 @@ TEST(test_json, memory_leak)
 
     EXPECT_EQ(2, object["key"].to_int());
 }
+
+TEST(test_json, contains)
+{
+    bourne::json object = {"key", 1, "nested", {"key", 2}};
+
+    EXPECT_TRUE(object.contains(object));
+    EXPECT_TRUE(object.contains(object["nested"]));
+    EXPECT_FALSE(object["nested"].contains(object));
+
+    bourne::json other = {"key", 1, "nested", {"key", 2}};
+
+    EXPECT_TRUE(other == object); // Check that the objects are equal
+    EXPECT_FALSE(
+        object.contains(other)); // Check that the objects are not contained in
+                                 // each other even though they are equal
+}