+++ NEWS.rst
@@ -10,6 +10,8 @@ Latest
   cause a memory leak.
 * Patch: Fixed bug where re-assignment via the ``operator[]`` would cause a memory
   leak.
+* Minor: Added ``json::contains`` function for checking if a json object is contained
+  in another json object.
 
 10.0.1
 ------